1_phidata_finance_agent/requirements.txt
phi-ai
python-dotenv
groq
openai
yfinance
duckduckgo-search
